+++ libs/dot-components/src/lib/components/empty-state/EmptyState.tsx
@@ -1,11 +1,25 @@
 import React from 'react';
+import styled from 'styled-components';
+import { Typography } from '@material-ui/core';
 import { ButtonProps, DotButton } from '../button/Button';
+import { CommonProps } from '../CommonProps';
+import { useStylesWithRootClass } from '../makeStylesWithRootClass';
 
-import './EmptyState.scss';
+const StyledEmptyState = styled.div`
+  &.dot-empty-state {
+    margin: 0 auto;
+    max-width: 600px;
+    text-align: center;
 
-export interface EmptyStateProps {
+    .empty-state-image {
+      min-height: 239px;
+    }
+  }
+`;
+
+export interface EmptyStateProps extends CommonProps {
   buttonProps?: ButtonProps;
-  Image?: React.FunctionComponent<
+  image?: React.FunctionComponent<
     React.SVGProps<SVGSVGElement> & { title?: string }
   >;
   imageAltText?: string;
@@ -13,29 +27,30 @@ export interface EmptyStateProps {
   title: string;
 }
 
-/**
- * @experimental This component is still in development
- */
 export const DotEmptyState = ({
   buttonProps,
-  Image,
+  className,
+  'data-testid': dataTestId,
+  image: Image,
   imageAltText,
-  title,
   subtitle,
+  title,
 }: EmptyStateProps) => {
+  const rootClasses = useStylesWithRootClass('dot-empty-state', className);
+
   return (
-    <div className="dot-empty-state empty-state-container">
+    <StyledEmptyState className={rootClasses} data-testid={dataTestId}>
       {Image && (
         <Image
-          title={imageAltText || title}
           className="empty-state-image"
           role="img"
+          title={imageAltText || title}
         />
       )}
-      <h4>{title}</h4>
-      {subtitle && <p>{subtitle}</p>}
+      <Typography variant="h2">{title}</Typography>
+      {subtitle && <Typography variant="body1">{subtitle}</Typography>}
       {buttonProps && <DotButton {...buttonProps} />}
-    </div>
+    </StyledEmptyState>
   );
 };